aula_02_titanic_kmeans/kmeans_function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
SQLAlchemy
pandas
psycopg2